Collect exceptions under `KrackenError` and move them to `errors.rb`
This is also intended to clean up some of the duplicate code that was a
by product of the changes living in an app before merged upstream.
