+++ spec/dummy/config/application.rb
@@ -9,6 +9,7 @@ require "sprockets/railtie"
 
 Bundler.require(*Rails.groups)
 require "kracken"
+require "kracken/json_api"
 
 module Dummy
   class Application < Rails::Application
@@ -23,6 +24,9 @@ module Dummy
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
+    config.middleware.insert_after ActionDispatch::DebugExceptions,
+                                   Kracken::JsonApi::PublicExceptions
+
   end
 end
